--- docs/source/api_reference/framework/time.rst
@@ -1,1 +1,0 @@
-.. automodule:: vivarium.framework.time
