+++ docs/source/api_reference/framework/time/index.rst
@@ -1,0 +1,11 @@
+===============
+Time Management
+===============
+
+.. automodule:: vivarium.framework.time
+
+.. toctree::
+   :maxdepth: 1
+   :glob:
+
+   *
