+++ docs/source/api_reference/framework/time/interface.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium.framework.time.interface
