+++ docs/source/api_reference/framework/time/manager.rst
@@ -1,0 +1,1 @@
+.. automodule:: vivarium.framework.time.manager
