+++ src/vivarium/framework/time/manager.py
@@ -1,7 +1,7 @@
 """
-====================
-The Simulation Clock
-====================
+============
+Time Manager
+============
 
 The components here provide implementations of different kinds of simulation
 clocks for use in ``vivarium``.
@@ -14,7 +14,6 @@ For more information about time in the simulation, see the associated
 from __future__ import annotations
 
 import math
-from collections.abc import Callable, Sequence
 from functools import partial
 from typing import TYPE_CHECKING, Any
 
@@ -28,11 +27,10 @@ if TYPE_CHECKING:
     from vivarium.framework.engine import Builder
     from vivarium.framework.event import Event
     from vivarium.framework.population import SimulantData
-    from vivarium.framework.resource import Resource
     from vivarium.framework.values import ValuesManager
 
 from vivarium.framework.values import list_combiner
-from vivarium.manager import Interface, Manager
+from vivarium.manager import Manager
 
 
 class SimulationClock(Manager):
@@ -300,52 +298,3 @@ class DateTimeClock(SimulationClock):
 
     def __repr__(self) -> str:
         return "DateTimeClock()"
-
-
-class TimeInterface(Interface):
-    """Public interface for the simulation time management system."""
-
-    def __init__(self, manager: SimulationClock) -> None:
-        self._manager = manager
-
-    def clock(self) -> Callable[[], ClockTime]:
-        """Gets a callable that returns the current simulation time."""
-        return lambda: self._manager.time
-
-    def step_size(self) -> Callable[[], ClockStepSize]:
-        """Gets a callable that returns the current simulation step size."""
-        return lambda: self._manager.step_size
-
-    def simulant_next_event_times(self) -> Callable[[pd.Index[int]], pd.Series[ClockTime]]:
-        """Gets a callable that returns the next event times for simulants."""
-        return self._manager.simulant_next_event_times
-
-    def simulant_step_sizes(self) -> Callable[[pd.Index[int]], pd.Series[ClockStepSize]]:
-        """Gets a callable that returns the simulant step sizes."""
-        return self._manager.simulant_step_sizes
-
-    def move_simulants_to_end(self) -> Callable[[pd.Index[int]], None]:
-        """Gets a callable that moves simulants to the end of the simulation"""
-        return self._manager.move_simulants_to_end
-
-    def register_step_size_modifier(
-        self,
-        modifier: Callable[[pd.Index[int]], pd.Series[ClockStepSize]],
-        required_resources: Sequence[str | Resource] = (),
-    ) -> None:
-        """Registers a step size modifier.
-
-        Parameters
-        ----------
-        modifier
-            Modifier of the step size pipeline. Modifiers can take an index
-            and should return a series of step sizes.
-        required_resources
-            A list of resources that need to be properly sourced before the
-            pipeline source is called. This is a list of strings, pipelines,
-            or randomness streams.
-
-        """
-        return self._manager.register_step_modifier(
-            modifier=modifier, required_resources=required_resources
-        )
